Ok, let's do this!
Finally figured out how everything works using a different assignment as reference. Learned that sandboxes basically seperate different segments of the page to make it easier to organize and to change the grid's location (useful for working with different parts of the body). Also have my references for color and different javascript shapes so now it's time to draw the character! Woohoo!